drop support for java 1.6/java6se
According to Wikipedia[0] Java SE 6 was released in December 2006 and
reached end-of-life in February 2013. Since then multiple security
vulnerabilities were fixed but those fixes are not available publicly.
As we are writing security-related software this is a no-go.

In addition parts of our toolchain (findbugs) are not working with
Java6 anymore.

This drops support for Java6, Java7 and 8 are supported as usual. I
will not add patches that were written just for the sake of utilizing
Java7+ features but won't decline patches that are using Java7-Features.

At least the travisci build will now finish successfully.

[0] http://en.wikipedia.org/wiki/Java_version_history