restructure code and introduce CryptoEngine
The algorithms are now split between at.archistar.crypto.secretsharing
and at.archistar.crypto.informationchecking. As it gets tedious to
configure the algorithms a new interface CryptoEngine was added. The
basic idea is that end-users are just instantiating a CryptoEngine and
the cryptoengines themselve configure all needed compontents (like
symmetric encryption algorithms, secret-sharing, rng,
information-checking, decoder). This resembles Spring a little bit.

Let's start with an RabinBenOrEngine and a CevalloEngine.