move GF-math from static into instance methods
We will need this to test some algorithms within GF(257).

A problem is that Java does not support static method within interfaces.
This prevents us from creating a clean abstraction for the fields that
we need for our mathematics stuff. The (architectural) solution is to
move this math functionality from static class methods into instance
methods.

This leads to a performance hit of ~30%. Which is a lot. But then this
library is mostly thought to be used for research purposes where
relative performance can be seen as to be of more performance than
absolute performance. We already think about creating a "new" library
that only offeres the most viable algorithms which then will also be
optimized. It's easier to optimize well-written code than doing stuff
in the other direction.